requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
paho-mqtt==2.2.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0 